// web/src/generated/api/models/notificationKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type NotificationKind = typeof NotificationKind[keyof typeof NotificationKind];


export const NotificationKind = {
  BOOKING_CONFIRMED: 'BOOKING_CONFIRMED',
  BOOKING_REQUESTED: 'BOOKING_REQUESTED',
  BOOKING_RESCHEDULED: 'BOOKING_RESCHEDULED',
  BOOKING_CANCELLED: 'BOOKING_CANCELLED',
  BOOKING_REMINDER: 'BOOKING_REMINDER',
} as const;